feat: export `uibeam::escape` as a utility
